+++ lib/galaxy/webapps/galaxy/api/dynamic_actions.py
@@ -1,47 +1,6 @@
-import logging
-
 from galaxy.web import _future_expose_api_anonymous_and_sessionless as expose_api_anonymous_and_sessionless
 from galaxy.web.base.controller import BaseAPIController
-from galaxy.util import docstring_trim
-
-log = logging.getLogger(__name__)
-
-
-def get_dynamic_post_processing_actions(config_post_actions):
-    """ Extract information about all post processing actions from the galaxy.ini config file. """
-    actions = list()
-    for action_name in config_post_actions:
-        if ":" in action_name:
-            # Should be a submodule of actions (e.g. examples:microscope_control)
-            (module_name, class_name) = action_name.rsplit(":", 1)
-            module_name = 'galaxy.jobs.actions.dynamic.%s' % module_name.strip()
-            module = __import__(module_name, globals(),
-                                fromlist=['temp_module'])
-            class_object = getattr(module, class_name.strip())
-
-            print(class_object, module)
-
-        else:
-            # No module found it has to be explicitly imported.
-            module = __import__('galaxy.jobs.actions.dynamic',
-                                globals(), fromlist=['temp_module'])
-            class_object = getattr(globals(), action_name.strip())
-
-        doc_string = docstring_trim(class_object.__doc__)
-        split = doc_string.split('\n\n')
-        if split:
-            sdesc = split[0].strip()
-        else:
-            log.error(
-                'No description specified in the __doc__ string for %s.' % action_name)
-        if len(split) > 1:
-            description = split[1].strip()
-        else:
-            description = ''
-
-        actions.append(dict(actionpath=action_name,
-                            short_desc=sdesc, desc=description, cl=class_object))
-    return actions
+from galaxy.jobs.actions.post import get_dynamic_post_processing_actions
 
 
 class DynamicActionsController(BaseAPIController):
